# docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: iTodo API
  version: v1
  description: Cross-platform TodoList HTTP API for mini-program, web, and mobile clients.
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/todos:
    get:
      tags: [Todos]
      summary: Query todos
      parameters:
        - name: listId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, COMPLETED]
        - name: important
          in: query
          schema:
            type: boolean
        - name: myDay
          in: query
          schema:
            type: boolean
        - name: dueFrom
          in: query
          schema:
            type: string
            format: date
        - name: dueTo
          in: query
          schema:
            type: string
            format: date
        - name: remindFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: remindTo
          in: query
          schema:
            type: string
            format: date-time
        - name: keyword
          in: query
          schema:
            type: string
        - name: tagId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          $ref: '#/components/responses/TodoPageSuccess'
    post:
      tags: [Todos]
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
  /api/v1/todos/reorder:
    post:
      tags: [Todos]
      summary: Reorder todos in a list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderTodosRequest'
      responses:
        '200':
          description: Reordered
  /api/v1/todos/{todoId}:
    get:
      tags: [Todos]
      summary: Get a todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
    patch:
      tags: [Todos]
      summary: Update a todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
    delete:
      tags: [Todos]
      summary: Delete a todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          description: Deleted
  /api/v1/todos/{todoId}/complete:
    post:
      tags: [Todos]
      summary: Complete a todo
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
  /api/v1/todos/{todoId}/uncomplete:
    post:
      tags: [Todos]
      summary: Restore a completed todo to active
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
  /api/v1/todos/{todoId}/important:
    post:
      tags: [Todos]
      summary: Mark a todo as important
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
    delete:
      tags: [Todos]
      summary: Remove important mark
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
  /api/v1/todos/{todoId}/my-day:
    post:
      tags: [Todos]
      summary: Add a todo to My Day
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
    delete:
      tags: [Todos]
      summary: Remove a todo from My Day
      parameters:
        - $ref: '#/components/parameters/TodoId'
      responses:
        '200':
          $ref: '#/components/responses/TodoSuccess'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TodoId:
      name: todoId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
...
    TodoSuccess:
      description: Todo operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TodoResponseEnvelope'
    TodoPageSuccess:
      description: Todo page query succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TodoPageResponseEnvelope'
...
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
...
    ReorderItemRequest:
      type: object
      required: [id, sortOrder]
      properties:
        id: { type: string, format: uuid }
        sortOrder: { type: integer }
...
    ReorderTodosRequest:
      type: object
      required: [listId, items]
      properties:
        listId: { type: string, format: uuid }
        items:
          type: array
          items: { $ref: '#/components/schemas/ReorderItemRequest' }
    CreateTodoRequest:
      type: object
      required: [title]
      properties:
        listId: { type: string, format: uuid }
        title: { type: string, maxLength: 255 }
        note: { type: string }
        dueDate: { type: string, format: date }
        remindAt: { type: string, format: date-time }
        repeatRule: { type: string, maxLength: 255 }
        myDay: { type: boolean }
        important: { type: boolean }
        sortOrder: { type: integer }
    UpdateTodoRequest:
      type: object
      properties:
        listId: { type: string, format: uuid }
        title: { type: string, maxLength: 255 }
        note: { type: string }
        dueDate: { type: string, format: date }
        remindAt: { type: string, format: date-time }
        repeatRule: { type: string, maxLength: 255 }
        myDay: { type: boolean }
        important: { type: boolean }
        sortOrder: { type: integer }
        clearNote: { type: boolean }
        clearDueDate: { type: boolean }
        clearRemindAt: { type: boolean }
        clearRepeatRule: { type: boolean }
    TodoResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        listId: { type: string, format: uuid }
        title: { type: string }
        note: { type: string }
        status: { type: string, enum: [ACTIVE, COMPLETED] }
        importance: { type: string, enum: [NORMAL, IMPORTANT] }
        dueDate: { type: string, format: date }
        remindAt: { type: string, format: date-time }
        repeatRule: { type: string }
        completedAt: { type: string, format: date-time }
        sortOrder: { type: integer }
        myDay: { type: boolean }
        version: { type: integer, format: int64 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TodoResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data: { $ref: '#/components/schemas/TodoResponse' }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
    TodoPageResponseEnvelope:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/TodoResponse' }
            total: { type: integer, format: int64 }
            page: { type: integer, format: int64 }
            size: { type: integer, format: int64 }
        error: { $ref: '#/components/schemas/ApiError' }
        traceId: { type: string }
...